+++ packages/api-derive/src/elections/info.ts
@@ -8,6 +8,7 @@ import type { ITuple } from '@polkadot/types/types';
 import type { Observable } from '@polkadot/x-rxjs';
 import type { DeriveElectionsInfo } from './types';
 
+import { combineLatest, of } from '@polkadot/x-rxjs';
 import { map } from '@polkadot/x-rxjs/operators';
 
 import { memo } from '../util';
@@ -16,6 +17,14 @@ import { memo } from '../util';
 type Member = SeatHolder | ITuple<[AccountId, Balance]>;
 
 type Candidate = AccountId | ITuple<[AccountId, Balance]>;
+
+interface Constants {
+  candidacyBond: Balance;
+  desiredRunnersUp: u32;
+  desiredSeats: u32;
+  termDuration: BlockNumber;
+  votingBond: Balance;
+}
 
 function isSeatHolder (value: Member): value is SeatHolder {
   return !Array.isArray(value);
@@ -41,31 +50,55 @@ function sortAccounts ([, balanceA]: [AccountId, Balance], [, balanceB]: [Accoun
   return balanceB.cmp(balanceA);
 }
 
+function getConstants (api: ApiInterfaceRx, elections: string): Constants {
+  return api.consts[elections]
+    ? {
+      candidacyBond: api.consts[elections].candidacyBond as Balance,
+      desiredRunnersUp: api.consts[elections].desiredRunnersUp as u32,
+      desiredSeats: api.consts[elections].desiredMembers as u32,
+      termDuration: api.consts[elections].termDuration as BlockNumber,
+      votingBond: api.consts[elections].votingBond as Balance
+    }
+    : {
+      candidacyBond: api.registry.createType('Balance'),
+      desiredRunnersUp: api.registry.createType('u32'),
+      desiredSeats: api.registry.createType('u32'),
+      termDuration: api.registry.createType('BlockNumber'),
+      votingBond: api.registry.createType('Balance')
+    };
+}
+
 function queryElections (api: ApiInterfaceRx): Observable<DeriveElectionsInfo> {
-  const section = api.query.phragmenElection
+  const elections = api.query.phragmenElection
     ? 'phragmenElection'
     : api.query.electionsPhragmen
       ? 'electionsPhragmen'
       : 'elections';
+  const [council] = api.registry.getModuleInstances(api.runtimeVersion.specName.toString(), 'council') || ['council'];
 
-  return api.queryMulti<[Vec<AccountId>, Vec<Candidate>, Vec<Member>, Vec<Member>]>([
-    api.query.council.members,
-    api.query[section].candidates,
-    api.query[section].members,
-    api.query[section].runnersUp
-  ]).pipe(
+  return (
+    api.query[elections]
+      ? api.queryMulti<[Vec<AccountId>, Vec<Candidate>, Vec<Member>, Vec<Member>]>([
+        api.query[council].members,
+        api.query[elections].candidates,
+        api.query[elections].members,
+        api.query[elections].runnersUp
+      ])
+      : combineLatest<[Vec<AccountId>, Candidate[], Member[], Member[]]>([
+        api.query[council].members,
+        of([]),
+        of([]),
+        of([])
+      ])
+  ).pipe(
     map(([councilMembers, candidates, members, runnersUp]): DeriveElectionsInfo => ({
-      candidacyBond: api.consts[section].candidacyBond as Balance,
+      ...getConstants(api, elections),
       candidateCount: api.registry.createType('u32', candidates.length),
       candidates: candidates.map(getCandidate),
-      desiredRunnersUp: api.consts[section].desiredRunnersUp as u32,
-      desiredSeats: api.consts[section].desiredMembers as u32,
       members: members.length
         ? members.map(getAccountTuple).sort(sortAccounts)
         : councilMembers.map((accountId): [AccountId, Balance] => [accountId, api.registry.createType('Balance')]),
-      runnersUp: runnersUp.map(getAccountTuple).sort(sortAccounts),
-      termDuration: api.consts[section].termDuration as BlockNumber,
-      votingBond: api.consts[section].votingBond as Balance
+      runnersUp: runnersUp.map(getAccountTuple).sort(sortAccounts)
     }))
   );
 }