+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Allow `instances` API config to drive collective locations
+
+
 ## 4.16.2 Jun 27, 2021
 
 Upgrade priority: Low. Recommended for Polkadot with runtime >= 9050.